refactor: reorganize deprecated files
