+++ .changeset/witty-avocados-pump.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-common': patch
+---
+
+Deprecate the legacy error and request logging handler types, respectively: `ErrorHandlerOptions` and `RequestLoggingHandlerFactory`.
