+++ packages/backend-common/src/deprecated/logging/index.ts
@@ -14,4 +14,9 @@
  * limitations under the License.
  */
 
-export { loggerToWinstonLogger } from './loggerToWinstonLogger';
+export { getRootLogger, getVoidLogger, setRootLogger } from './globalLoggers';
+export {
+  createRootLogger,
+  redactWinstonLogLine,
+  coloredFormat,
+} from './createRootLogger';
