+++ packages/backend-common/src/deprecated/middleware/notFoundHandler.ts
@@ -15,23 +15,24 @@
  */
 
 // eslint-disable-next-line @backstage/no-relative-monorepo-imports
-import { MiddlewareFactory } from '../../../backend-app-api/src/http/MiddlewareFactory';
-import { RequestHandler } from 'express';
-import { LoggerService } from '@backstage/backend-plugin-api';
-import { getRootLogger } from '../logging';
+import { MiddlewareFactory } from '../../../../backend-app-api/src/http/MiddlewareFactory';
 import { ConfigReader } from '@backstage/config';
+import { RequestHandler } from 'express';
+import { getRootLogger } from '../logging';
 
 /**
- * Logs incoming requests.
+ * Express middleware to handle requests for missing routes.
+ *
+ * Should be used as the very last handler in the chain, as it unconditionally
+ * returns a 404 status.
  *
  * @public
- * @param logger - An optional logger to use. If not specified, the root logger will be used.
  * @returns An Express request handler
- * @deprecated @deprecated Use {@link @backstage/backend-app-api#MiddlewareFactory.create.logging} instead
+ * @deprecated Use {@link @backstage/backend-app-api#MiddlewareFactory.create.notFound} instead
  */
-export function requestLoggingHandler(logger?: LoggerService): RequestHandler {
+export function notFoundHandler(): RequestHandler {
   return MiddlewareFactory.create({
     config: new ConfigReader({}),
-    logger: logger ?? getRootLogger(),
-  }).logging();
+    logger: getRootLogger(),
+  }).notFound();
 }
