+++ packages/backend-common/src/middleware/index.ts
@@ -14,7 +14,4 @@
  * limitations under the License.
  */
 
-export * from './errorHandler';
-export * from './notFoundHandler';
-export * from './requestLoggingHandler';
 export * from './statusCheckHandler';
