+++ packages/backend-common/src/service/index.ts
@@ -14,6 +14,4 @@
  * limitations under the License.
  */
 
-export { createServiceBuilder } from './createServiceBuilder';
 export { createStatusCheckRouter } from './createStatusCheckRouter';
-export type { ServiceBuilder, RequestLoggingHandlerFactory } from './types';
